requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.14
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary~=2.9.6
pydantic==1.10.10
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
